Perform type conversion in pair() function
This commit changes the pair() function to perform dynamic type
conversion on the provided arguments.

This is needed because some private key files contain both private and
public key information, so

    read my.key

will return a KeyPair

hence:

    pair ( read my.key, read my.crt )

will attempt to construct a key pair from (KeyPair, Certificate) which
requires conversion of the KeyPair argument to a PrivateKey.