Support recursive set-password on a keystore
A PKCS#12 (or JKS) keystore has a password for the keystore itself as
well as a password for each private key entry. Typically, these
passwords are the same, but that is not required.

If a keystore uses a common password for the stores and the keys, and
you change the password on the keystore _only_, then it will break the
expectation that the keystore and the keys share a password.

This commit changes the `set-password` command to add a `-r` option so
that changing the password in the keystore can also change the
password for any keys that had the same password as the store.