Replace valueDescription w/ toTextValue in X5Value
The X5Value interface has a "toTextValue" method by virtue of
extending the "ToTextValue" interface. The default implementation
provides a textual representation of the X5Value's underlying "value()"

AbstractValueType also had a "valueDescription" method which was used
as part of the default implementation of "description()" and
duplicated the functionality of "toTextValue".

This commit removes "valueDescription()" and uses "toTextValue()" in
its place so that derived classes only need to override one method in
order to get appropriate output everywhere.